+++ src/app/views/movie-detail/client-side.tsx
@@ -4,7 +4,7 @@ import MovieDetailBody from '@/app/views/movie-detail/components/organisms/body'
 import MovieDetailHead from '@/app/views/movie-detail/components/organisms/head';
 import { IMovieDetailProps } from '@/app/views/movie-detail/interface';
 import ClientSideBuilder from '@/library/modules/page-generator/builder/client-side.builder';
-import { SEOHTMLGenerator } from '@/library/modules/seo/helper/seo-html.helper';
+import { SEOHTMLGenerator } from '@/library/modules/seo/helper';
 
 /**
  * Movie Detail
