+++ src/app/views/movie-detail/index.ts
@@ -1,0 +1,2 @@
+export { getServerSideProps } from '@/app/views/movie-detail/server-side';
+export { default } from '@/app/views/movie-detail/client-side';
