+++ src/library/abstract/interface/index.ts
@@ -1,0 +1,1 @@
+export * from './builder.interface';
