+++ src/library/interface/general/base.interface.ts
@@ -10,7 +10,6 @@ import { ISeoProps } from './seo.interface';
 export interface IBaseProps {
   csrfToken: string;
   language: ILanguagePreset;
-  withRedux: boolean;
   // TODO: will add aditional props
 }
