+++ src/library/interface/language/index.ts
@@ -1,0 +1,1 @@
+export * from './language.interface';
