+++ src/library/interface/movie/index.ts
@@ -1,0 +1,1 @@
+export * from './movie-type.interface';
