+++ src/library/model/general/interface/index.ts
@@ -1,0 +1,3 @@
+export * from './pagination.interface';
+export * from './rating.interface';
+export * from './score.interface';
