+++ src/library/model/movie/interface/index.ts
@@ -1,0 +1,3 @@
+export * from './movie-detail.interface';
+export * from './movie-figure.interface';
+export * from './movie-list.interface';
