+++ src/library/modules/cors/helper/index.ts
@@ -1,0 +1,1 @@
+export * from './cors-generator.helper';
