+++ src/library/modules/cors/interface/index.ts
@@ -1,0 +1,1 @@
+export * from './cors.interface';
