+++ src/library/modules/error/helper/index.ts
@@ -1,0 +1,1 @@
+export * from './error-apps.helper';
