+++ src/library/modules/error/interface/index.ts
@@ -1,0 +1,1 @@
+export * from './error.interface';
