+++ src/library/modules/fetch-api/interface/index.ts
@@ -1,0 +1,1 @@
+export * from './fetch-api.interface';
