+++ src/library/modules/next-api/interface/index.ts
@@ -1,0 +1,1 @@
+export * from './next-api.interface';
