+++ src/library/modules/page-generator/helper/index.ts
@@ -1,0 +1,1 @@
+export * from './page-server-side.helper';
