+++ src/library/modules/seo/helper/index.ts
@@ -1,0 +1,1 @@
+export * from './seo-html.helper';
