+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import App, { AppInitialProps, AppProps } from 'next/app';
+import { AppProps } from 'next/app';
 
 import '@/library/styles/css/global.css';
 
@@ -10,17 +10,5 @@ import '@/library/styles/css/global.css';
 const MyApp = ({ Component, pageProps }: AppProps) => (
   <Component {...pageProps} />
 );
-
-MyApp.getInitialProps = async (appContext): Promise<AppInitialProps> => {
-  const appProps = await App.getInitialProps(appContext);
-
-  return {
-    ...appProps,
-    pageProps: {
-      ...appProps.pageProps,
-      isLoggedIn: true
-    }
-  };
-};
 
 export default MyApp;
